Improve interactions (#11)
* move interactions code to respective file

* improve interactions semantics: 'name:target:params'

* make dbus calls async

* avoid deadlocks from dbus race conditions

The tutorial main loop and other components are simultaneously
dbus listeners and emitters. If a dbus event triggers back another
event on the other component, both components can get stuck waiting
on eachother's responses.

To solve this, we ignore the replys made by coponent -> tutorial

* fix unresponsive dbus calls

The components like the watchers were sending dbus messages to the
tutorial main loop via interactions.register() method. However,
these were not being sent on the bus.

Introspection turned out to be the source of the issue. For some
reason it never return from the introspection and thus never called
the method. Disabling introspection seems to solve the issue.

* revert async calls for qubes-menu

All calls are now syncronous. The difference is that some ignore
the replies

* small fixes