make TutorialUIInterface inherit from Gtk.Window
Previously, all UI components were inheriting from Gtk.Window
and TutorialUIInterface.

There was a lot of code repetition since we had not yet found a way
to inheirit from a subclass of Gtk.Window. Now that's been solved.

The solution is not perfect, though. When editing with glade, it
requires us to remove some comment and then re-add it. But the code
will be more reabable as a tradeoff.